Modifiy nginx.tmpl to allow both 443 and 80.
Allow http when https is activated. This parameter avoid infinite redirect when the reverse proxy [nginx-proxy](https://github.com/jwilder/nginx-proxy) is already behind another reverse proxy.

This parameter is not recommended for production. Use it for test / transition.

To use it, launch your container with the environnement variable `ALLOW_80=1` :

    $ docker run -e VIRTUAL_HOST=foo.bar.com  ...