Add useful debugging comments into the generated nginx config file
In the issue tracker, as many share their generated nginx config without context, it is impossible to get a glimpse of what could be wrong without starting a conversation asking the OP for more details/context.

With this changes, the generated nginx config file will contain useful comments about environment variables, networks that were set at the time of generation.