nginx.tmpl: use closest vhost.d and htpasswd files (#1309)
Add logic to find the closest matching files for vhost.d, vhost.d
_include and htpasswd files.

This allows multiple hosts to share these files, similar to wildcard
certificate files, but not limited to just 1 subdomain depth.

The match is anchored to the end of the string filename, so a largest
suffix match is found.

For example, 'vhost.d/b.c' will match a 'a.b.c' virtual host.
But 'a.b' will not match 'a.b.c'.
This is stricter than the "closest" match function that certificate
files used in the past, which would also match in the 2nd case.

This is a WIP, unfinished sections are marked with TODO.